Imports: move merge_yaml to spack.schema
