+++ lib/spack/spack/ci/common.py
@@ -26,6 +26,7 @@ import spack.deptypes as dt
 import spack.environment as ev
 import spack.error
 import spack.mirrors.mirror
+import spack.schema
 import spack.spec
 import spack.util.spack_yaml as syaml
 import spack.util.url as url_util
@@ -608,7 +609,7 @@ class SpackCIConfig:
                     if "build-job-remove" in match_attrs:
                         spack.config.remove_yaml(dest, attrs["build-job-remove"])
                     if "build-job" in match_attrs:
-                        spack.config.merge_yaml(dest, attrs["build-job"])
+                        spack.schema.merge_yaml(dest, attrs["build-job"])
                     break
             if matched and only_first:
                 break
@@ -679,7 +680,7 @@ class SpackCIConfig:
                     if do_remove:
                         dest = spack.config.remove_yaml(dest, src[remove_job_name])
                     if do_merge:
-                        dest = copy.copy(spack.config.merge_yaml(dest, src[merge_job_name]))
+                        dest = copy.copy(spack.schema.merge_yaml(dest, src[merge_job_name]))
 
                 if name == "build":
                     # Apply attributes to all build jobs
@@ -808,7 +809,7 @@ class SpackCIConfig:
                             tty.warn(f"Response missing required keys: {missing_keys}")
 
                     if clean_config:
-                        job["attributes"] = spack.config.merge_yaml(
+                        job["attributes"] = spack.schema.merge_yaml(
                             job.get("attributes", {}), clean_config
                         )
